big WIP commit for last chunk of overhauled infrastructure
Finally ready to start using these new tools to work with Lecroy files
