+++ setup.py
@@ -9,10 +9,10 @@ setupkwargs = dict(
     packages=setuptools.find_packages(include=['openmsipython*']),
     include_package_data=True,
     entry_points = {
-        'console_scripts' : ['manage_service=openmsipython.command_line_scripts.manage_service:main',
-                             'upload_data_files_added_to_directory=openmsipython.command_line_scripts.upload_data_files_added_to_directory:main',
-                             'upload_data_file=openmsipython.command_line_scripts.upload_data_file:main',
-                             'reconstruct_data_files=openmsipython.command_line_scripts.reconstruct_data_files:main',
+        'console_scripts' : ['upload_data_file=openmsipython.data_file_io.upload_data_file:main',
+                             'data_file_upload_directory=openmsipython.data_file_io.data_file_upload_directory:main',
+                             'data_file_download_directory=openmsipython.data_file_io.data_file_download_directory:main',
+                             'manage_service=openmsipython.services.manage_service:main',
                             ],
     },
     python_requires='>=3.7,<3.8',
