+++ test/unittests/testdatafilechunk_with_kafka.py
@@ -1,6 +1,6 @@
 #imports
 from config import TEST_CONST
-from openmsipython.data_file_io.data_file import UploadDataFile
+from openmsipython.data_file_io.upload_data_file import UploadDataFile
 from openmsipython.data_file_io.data_file_chunk import DataFileChunk
 from openmsipython.data_file_io.config import RUN_OPT_CONST
 from openmsipython.my_kafka.my_producers import MySerializingProducer
